+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Allow runtime update of Sidecar images
 - Allow Agent recreation with preserved IDs
 - The internal metrics exporter can not be disabled
+- Allow to customize SchedulerName inside Member Pod
 
 ## [1.2.2](https://github.com/arangodb/kube-arangodb/tree/1.2.2) (2021-09-09)
 - Update 'github.com/arangodb/arangosync-client' dependency to v0.7.0